+++ tests/rule_tests/mod.rs
@@ -223,6 +223,7 @@ mod remove_compound_assignment;
 mod remove_empty_do;
 mod remove_method_definition;
 mod remove_nil_declaration;
+mod remove_types;
 mod remove_unused_if_branch;
 mod remove_unused_while;
 mod rename_variables;